Refactoring.  Material is now purely an interface, generating no symbols except for a default constructor and a virtual destructor, both of which are defined in the .h.  material.cpp therefore isn't needed; all the logic there is moved to solidmaterial.cpp where it belongs.  A default material was being created in a few places; they have been replaced with SolidMaterial but probably long term need to be revisited to determine if they make sense.
Also changed show-sample target in makefile to use gtksee, which is a clone of an image viewer of similair name.  I'm still not thrilled with this choice, but it's the best I've found in the (limited) amount of time I spent searching.



git-svn-id: https://svn.code.sf.net/p/mbrt/code/trunk@50 a272ff57-b438-0410-9c92-d377374fc898

+++ plugins/noisematerial.cpp
@@ -98,8 +98,8 @@ double NoiseMaterial::choose_material(const Point3D& intersection_point) const
     d0 = p10 - p00;
     d1 = p11 - p01;
 
-    p0 = floor(d0 * oy) + p00;
-    p1 = floor(d1 * oy) + p01;
+    p0 = (int) floor(d0 * oy) + p00;
+    p1 = (int) floor(d1 * oy) + p01;
 
     d = p1 - p0;
